#1882 Implemented support for test PopulateDb run against PosgreSQL.
