+++ platform-dao/src/test/java/ua/com/fielden/platform/test_config/PostgresqlDbCreator.java
@@ -20,6 +20,7 @@ import ua.com.fielden.platform.entity.query.metadata.PersistedEntityMetadata;
 import ua.com.fielden.platform.test.AbstractDomainDrivenTestCase;
 import ua.com.fielden.platform.test.DbCreator;
 import ua.com.fielden.platform.test.IDomainDrivenTestCaseConfiguration;
+import ua.com.fielden.platform.utils.DbUtils;
 
 /**
  * This is a DB creator implementation for running unit tests against PostgreSQL.
@@ -43,7 +44,7 @@ public class PostgresqlDbCreator extends DbCreator {
      */
     @Override
     protected List<String> genDdl(final DomainMetadata domainMetaData, final Dialect dialect) {
-        final List<String> result = PostgresqlDbUtils.prependDropDdlForPostgresql(domainMetaData.generateDatabaseDdl(dialect));
+        final List<String> result = DbUtils.prependDropDdlForPostgresql(domainMetaData.generateDatabaseDdl(dialect));
         // Drop all the foreign key constraints to allow out-of-order data truncation and/or population.
         // Need to pass the following PL/SQL as a single line otherwise it gets executed one independent line at a time, which does not work.
         result.add("do $$ declare r record; begin for r in (select table_name, constraint_name from information_schema.table_constraints where table_schema = 'public' and constraint_type = 'FOREIGN KEY') loop execute concat('alter table ' || r.table_name || ' drop constraint ' || r.constraint_name); end loop; end $$;");
