+++ platform-web-resources/src/main/resources/application.properties
@@ -22,30 +22,17 @@ web.domain=tgdev.com
 web.path=/
 
 # email settings
-email.smtp=mail.fielden.com.au
-email.fromAddress=trident.fleet@fielden.com.au
+email.smtp=localhost
+email.fromAddress=tg@locahost
 
 #################################################################
 #################### Hibernate settings #########################
 #################################################################
-#hibernate.connection.url=jdbc:jtds:sqlserver://192.168.1.3:4064/FP_AU
-#hibernate.connection.driver_class=net.sourceforge.jtds.jdbc.Driver
-#hibernate.dialect=org.hibernate.dialect.SQLServerDialect
-#hibernate.connection.username=t32
-#hibernate.connection.password=T32
-
-
-#hibernate.dialect=org.hibernate.dialect.SQLServer2012Dialect
-#hibernate.connection.driver_class=com.microsoft.sqlserver.jdbc.SQLServerDriver
-#hibernate.connection.url=jdbc:sqlserver://192.168.1.142:1433;database=TG_AIR_TEST_DB_1;queryTimeout=30
-#hibernate.connection.username=junit
-#hibernate.connection.password=junit
-
-hibernate.connection.url=jdbc:h2:./src/main/resources/db/devdb
-hibernate.connection.driver_class=org.h2.Driver
-hibernate.dialect=org.hibernate.dialect.H2Dialect
-hibernate.connection.username=sa
-hibernate.connection.password=
+hibernate.connection.url=jdbc:postgresql://localhost:5432/ci_1
+hibernate.connection.driver_class=org.postgresql.Driver
+hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
+hibernate.connection.username=junit
+hibernate.connection.password=junit
 
 ##############################
 ### Hikari connection pool ###
